requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
pathlib
plotly
matplotlib
openpyxl
pysimplegui